Lab 1, 2
Cleared Lab 1
Made lab 2

+++ Lab1/Proj6/Proj6.cpp
@@ -2,10 +2,10 @@
 
 int main()
 {
-	long long Текущий;
-	long long s = 60 * 60 * 24 * 365;
-	long long s100 = s * 100;
-	long long s1000 = s * 1000;
+	short Текущий;
+	short s = 60 * 60 * 24 * 365;
+	short s100 = s * 100;
+	short s1000 = s * 1000;
 	Текущий = 2021 * s;
 	return 0;
